requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0

# ── Gemini AI ──────────────────────────────────────────────────
google-generativeai==0.5.4

# ── Machine Learning ───────────────────────────────────────────
scikit-learn==1.5.2
pandas==2.2.2
numpy==2.4.4
joblib==1.4.2
